Clean up, move public api to its own module(s)
The public API needs to be duplicated for the unsafe_textures and
non-unsafe_textures versions, because of the `impl FontTexture<'r>`
lifetime annotation. So now the public API is split from the
implementation of creating the texture and drawing text, which don't
need any lifetime annotations for the texture (they're elided if
running without unsafe_textures).

Would be nice if the `impl FontTexture<'_>` lifetime could be elided
entirely (not just "_").